Move function collections model
